+++ src/v1/controllers/controllerCursos.js
@@ -5,6 +5,14 @@ const createCurso = async (req, res) => {
   try {
     const { curso } = req.body;
     if (!curso) return res.json({ message: "error" });
+
+    const respose = await prisma.curso.findFirst({
+      where: {
+        curso,
+      },
+    });
+    if (respose) return res.json({ message: "exist" });
+
     await prisma.curso.create({
       data: {
         curso,
@@ -56,12 +64,40 @@ const getCursoEspecifico = async (req, res) => {
 const deleteCurso = async (req, res) => {
   try {
     const { id } = req.params;
+
+    const response = await prisma.curso.findFirst({
+      include: {
+        cadeiraAtrasos: true,
+        exameEspecial: true,
+        recursos: true,
+        estudantes: true,
+        disciplinas: true,
+        inscricaoMatricula: true,
+      },
+      where: {
+        id,
+      },
+    });
+    if (
+      response.cadeiraAtrasos.length > 0 ||
+      response.exameEspecial.length > 0 ||
+      response.recursos.length > 0 ||
+      response.estudantes.length > 0 ||
+      response.disciplinas.length > 0 ||
+      response.inscricaoMatricula.length > 0
+    )
+      return res.json({
+        status: 400,
+        response: response,
+        message: "Não Podes Eliminar Este Curso Porque Tem uma Associação!",
+      });
+
     await prisma.curso.delete({
       where: {
         id,
       },
     });
-    res.json({ message: "Sucess" });
+    res.json({ message: "sucess" });
   } catch (error) {
     res.json({ message: "error" });
   }
@@ -84,6 +120,28 @@ const upDateCurso = async (req, res) => {
     res.json({ message: "error" });
   }
 };
+const cursoEspecifico = async (req, res) => {
+  try {
+    const { id } = req.params;
+    if (!id) return res.json({ message: "error" });
+    const response = await prisma.curso.findFirst({
+      include: {
+        cadeiraAtrasos: true,
+        exameEspecial: true,
+        recursos: true,
+        estudantes: true,
+        disciplinas: true,
+        inscricaoMatricula: true,
+      },
+      where: {
+        id,
+      },
+    });
+    res.json(response);
+  } catch (error) {
+    res.json({ message: "error" });
+  }
+};
 
 module.exports = {
   createCurso,
@@ -92,4 +150,5 @@ module.exports = {
   deleteCurso,
   upDateCurso,
   getCursoEspecifico,
+  cursoEspecifico,
 };
