+++ src/v1/controllers/controllerSemestre.js
@@ -7,6 +7,13 @@ const createSemestre = async (req, res) => {
     if ((!nome && !numero) || numero === 0)
       return res.json({ message: "Campo vazios" });
 
+    const respose = await prisma.semestre.findFirst({
+      where: {
+        nome,
+      },
+    });
+    if (respose) return res.json({ message: "exist" });
+
     await prisma.semestre.create({
       data: {
         nome,
@@ -27,13 +34,7 @@ const getSemestres = async (req, res) => {
     res.json(error);
   }
 };
-const SemestresEspecifico = async (req, res) => {
-  const { fk_curso, fk_ano } = req.body;
-  try {
-  } catch (error) {
-    res.json(error);
-  }
-};
+
 const getSemestre = async (req, res) => {
   try {
     const { id } = req.params;
@@ -64,6 +65,25 @@ const buscaSemestre = async (req, res) => {
 const deleteSemestre = async (req, res) => {
   try {
     const { id } = req.params;
+    const response = await prisma.semestre.findFirst({
+      include: {
+        cadeiraAtrasos: true,
+        exameEspecial: true,
+        recursos: true,
+        disciplinas: true,
+      },
+    });
+    if (
+      response.cadeiraAtrasos.length > 0 ||
+      response.exameEspecial.length > 0 ||
+      response.recursos.length > 0 ||
+      response.disciplinas.length > 0
+    )
+      return res.json({
+        status: 400,
+        message: "Não Podes Eliminar Este Semestres Porque Tem uma Associação!",
+      });
+
     await prisma.semestre.delete({
       where: {
         id,
@@ -93,6 +113,26 @@ const upDateSemestre = async (req, res) => {
     res.json({ message: "error" });
   }
 };
+const semestreEspecifico = async (req, res) => {
+  try {
+    const { id } = req.params;
+    if (!id) return res.json({ message: "error" });
+    const response = await prisma.semestre.findFirst({
+      include: {
+        cadeiraAtrasos: true,
+        exameEspecial: true,
+        recursos: true,
+        disciplinas: true,
+      },
+      where: {
+        id,
+      },
+    });
+    res.json(response);
+  } catch (error) {
+    res.json({ message: "error" });
+  }
+};
 
 module.exports = {
   createSemestre,
@@ -101,4 +141,5 @@ module.exports = {
   deleteSemestre,
   upDateSemestre,
   buscaSemestre,
+  semestreEspecifico,
 };
